Fixed issues with Gradle build, JDK, SDK, etc.
updated project to run off of updated software, updated gradle build to 8.7.3, jdk to 17, and SDK target to 34.  Minimum is 33.  Modified build file to also not have issues with the parse library which was pulling up firebase dispatcher which is deprecated.  Error is not firing when version is set to 4.3.0